ima: add functions to manage digest lists
This patch introduces a new structure called ima_digest, which includes a
digest, a set of flags, and the type of data the digest was calculated
from.

The patch also introduces functions to search and add a new digest to the
new ima_digests_htable hash table.

Changelog

v2:
- replace is_mutable with flags in the ima_digest structure definition
- add support for multiple hash algorithms
- add type to ima_digest structure

v1:
- add support for immutable/mutable files

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>