evm: add kernel command line option to select protected xattrs
The purpose of this patch is to select which extended attributes must be
protected by EVM and provide to IMA this information.

If a user requires that only security.ima must be protected (for example,
if the IMA policy does not rely on LSM labels), creating both security.ima
and security.evm for immutable files can be avoided when digest lists are
used as a source of reference values. The integrity of digest lists is
protected with a digital signature.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>