KEYS: PGP data parser
Implement a PGP data parser for the crypto key type to use when
instantiating a key.

This parser attempts to parse the instantiation data as a PGP packet
sequence (RFC 4880) and if it parses okay, attempts to extract a public-key
algorithm key or subkey from it.

If it finds such a key, it will set up a public_key subtype payload with
appropriate handler routines (DSA or RSA) and attach it to the key.

Thanks to Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp> for pointing
out some errors.

Changelog:

- do not process Public-Subkey packets (only one key will be created, from
  the Public-Key packet)

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>