ima: check if digest lists and metadata are measured/appraised
This patch checks if both digest lists and metadata are measured and if
digest lists are appraised. The flags of the actions for which digest
lookup is enabled are set in the new global variable
ima_digest_list_actions.

If the digest list or metadata measurements are missing, remote verifiers
cannot determine from the measurement list which files could have been
accessed. If the digital signature of digest lists is not verified, files
not signed by software vendors could be accessed.

Digest list lookup can be selectively enabled for a specific action (e.g.
measure, appraise) with the new 'ima_digest_list_actions=' kernel
parameter.

Changelog

v2:
- introduce ima_digest_list_check_action()
- remove hooks and policy
- add ima_digest_list_actions= kernel parameter

v1:
- clear IMA_MEASURE action flag only if it was in the policy
- check if at least one action is allowed before searching the file digest
- retrieve ima_digest structure
- set IMA action flags in ima_disable_digest_lookup
- add DIGEST_LIST_METADATA_CHECK hook
- update MEASURE/APPRAISE policies

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>