ima: allow the parser to read metadata without appraisal verification
The file containing digest list metadata cannot be provided by the vendor,
as its content depends on the set of packages installed in the system. If
that system only accepts files signed by the vendor, the parser couldn't
access the metadata.

Allow the parser to access metadata with the assumption that the parsing
code is reasonably robust against malformed data. This exception is granted
only to the parser and only when it opened the 'digest_list_metadata' file
in the securityfs filesystem.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>