ima: introduce securityfs interfaces for digest lists
This patch introduces 'digest_list_metadata' and 'digest_list_data' in the
securityfs filesystem to load respectively digest list metadata and digest
lists. Access to these files is granted only to the user space parser,
after the digest of the parser has been uploaded to IMA at boot time.

Initially, the parser uploads digest list metadata. IMA adds the digests of
digest lists to the hash table, after setting a flag to distinguish them
from digests of regular files.

Then, the parser accesses the original digest lists (whose format must be
supported by the parser), convert them to compact lists, and uploads the
data to IMA. IMA prevents a corrupted parser from uploading unexpected
digests by rejecting the write if the parser accessed files which are not
digest lists (the flag is not set in the digest found in the hash table).

The patch also introduces 'digests_count', to show the number of digests
stored in the ima_digests_htable hash table.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>

Changelog

v2:
- introduce two dentries to load digest list data and metadata
- call ima_parse_digest_list_metadata() and ima_parse_compact_list()
  from ima_write_data()
- remove ima_get_file_id()
- introduce functions to restrict access to new dentries by parser

v1:
- deny upload of digest lists if no policy is loaded