ima: added support for key data type
This patch allows users to load custom keys to be used for signature
verification of digest lists. Loading is permitted according to the IMA
keyring configuration: if CONFIG_INTEGRITY_TRUSTED_KEYRING is enabled, keys
can be loaded into the IMA keyring only if they are signed with a key in
the trusted keyring. If the option is not enabled, the user space parser
will first create the IMA keyring, before metadata is sent to IMA.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>