ima: execute parser at boot time and parse digest lists
At kernel initialization time, IMA directly loads the parser metadata from
a predefined location, and then calls the user mode helper to execute the
parser in user space.

Without file signatures, the digest of the parser (included in the
metadata) must be added to the hash table, after the signature is verified,
so that the parser can be executed when appraisal is in enforcing mode.

The parser reads the digest lists and uploads the extracted digests to IMA.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>